refactor: 장바구니 right, left section 분리
