+++ src/components/domain/index.ts
@@ -2,3 +2,5 @@ export { default as Product } from './Product';
 export { default as ProductCard } from './ProductCard';
 export { default as Cart } from './Cart';
 export { default as CartCard } from './CartCard';
+export { default as CartLeftSection } from './CartSection/CartLeftSection';
+export { default as CartRightSection } from './CartSection/CartRightSection';
